backend/pyproject.toml
[project]
name = "chen-assistant"
version = "0.1.0"
description = "here is my dream."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "jieba>=0.42.1",
    "langchain>=1.0",
    "langgraph>=1.0",
    "langchain-chroma>=0.2",
    "langchain-community>=0.3",
    "langchain-openai>=0.3",
    "pydantic-settings>=2.0",
    "chromadb>=1.0",
    "dashscope>=1.20",
    "docx2txt>=0.9",
    "pypdf>=5.0",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.27",
    "uvicorn>=0.46.0",
]
